src/06-destructuring-assignment/destructuring.tsx
import React from "react";

export type ManType = {
    name: string,
    age: number,
    lessons: Array<{ title: string }>,
    address: {
        street: {
            title: string
        }
    }
}

type PropsType = {
    title: string
    man: ManType
    food: Array<string>
    car: {model: string}
}

export const ManComponent: React.FC<PropsType> = ({title, man, ...restProps}) => {
  //  const {title, man: {name}} = props; //можно так

    return <div>
        <h1>{title}</h1>
        <hr/>
        <div>
            {man.name}
        </div>
        <div>
            {restProps.car.model}
        </div>
    </div>
}